move PropertyTypesContext provider out of definition tab
